rename Raft to Client
